requirements.txt
httplib2==0.9.2
isodate==0.5.4
jdcal==1.2
Jinja2==2.8
keyring==9.0
launchpadlib==1.10.3
MarkupSafe==0.23
paramiko==1.15.2
parsedatetime==1.5
pbr==1.8.1
python-dateutil==2.2
pytimeparse==1.1.5
pytz==2015.7
PyYAML==3.11
requests==2.7.0
simplejson==3.8.2
urllib3==1.13.1
